Store a declaration for every builtin declaration
Also, make TokenArray generic, rename it to ChunkedArray
