+++ src/parse.hpp
@@ -3,8 +3,7 @@
 #include <memory>
 
 #include "parser.hpp"
-
-struct TokenArray;
+#include "token_array.hpp"
 
 Writer<std::unique_ptr<AST::AST>> parse_program(std::string const&, TokenArray&);
 Writer<std::unique_ptr<AST::AST>> parse_expression(std::string const&, TokenArray&);
